Split Transactions to TransactionList component
